+++ src/parallel/mod.rs
@@ -1,4 +1,3 @@
-mod error;
 pub(crate) mod multi;
 mod ordered;
 pub(crate) mod paired;
@@ -8,7 +7,6 @@ mod processor;
 pub(crate) mod reader;
 pub(crate) mod single;
 
-pub use error::{IntoProcessError, ProcessError, Result};
 pub use ordered::Ordered;
 pub use processor::{MultiParallelProcessor, PairedParallelProcessor, ParallelProcessor};
 pub use reader::ParallelReader;