+++ src/prelude.rs
@@ -1,7 +1,7 @@
 pub use crate::{
     parallel::{
-        IntoProcessError, MultiParallelProcessor, PairedParallelProcessor, ParallelProcessor,
-        ParallelReader, PoolParallelReader, ThreadPool,
+        MultiParallelProcessor, PairedParallelProcessor, ParallelProcessor, ParallelReader,
+        PoolParallelReader, ThreadPool,
     },
-    Record,
+    IntoParaseqError, Record,
 };